k1c: Add management of ASN for the IOMMU
Summary:
This patch reads IOMMU ID from device tree and create a group and a
domain according to this ID. If two devices have the same ID then they
will be in the same group (and so in the same domain). Now in the device
tree you need to delcare:

        iommu_noc: iommu_noc@0 {
                compatible = "kalray,k1c-iommu";
		...
                #iommu-cells = <1>;
        };

        test_dmaengine@0 {
                compatible = "kalray,k1c-dma-noc-test";
		...
                iommus = <&iommu_noc 2>;
        };

        dma_noc: dma_noc@800000 {
                compatible = "kalray,k1c-dma-noc";
		...
                iommus = <&iommu_noc 2>;
        };

In this case test_dmaengine and dma_noc will be in the same group and
share the same ID.

Ref T9361

Test Plan: It can not be tested for now because it is not supported by devices.

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D1310