k1c: Add optimized clear_page
Summary:
Optimized clear_page by using dzerol instruction. dzerol is used to zero
cache lines directly. This is easy to do on a page since it is always
aligned. Moreover, use an hardware loop to execute this for all cache
lines matching the page.

Test Plan:
Checked assembly in a separate test without the kernel and tested
it in the kernel.

Reviewers: O51 Linux Coolidge, gthouvenin

Reviewed By: O51 Linux Coolidge, gthouvenin

Subscribers: ysionneau, alfred, #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D1306