k1c: Move huge page handling in slow path
Summary:
Huge path are less refill-proned and thus, we don't want to impact the
classic refill when used. Best solution is to move them out of fast
path and handle them in a slow path. Moreover,  rework the refill part
after modifications for huge page to reduce the number of operations
needed to refill.

Test Plan: Checked that coolval with huge pages still runs correctly.

Reviewers: O51 Linux Coolidge, gthouvenin

Reviewed By: O51 Linux Coolidge, gthouvenin

Subscribers: gthouvenin, alfred, #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D1263