k1c: support an already initialized IOMMU
Summary:
This patch allows to probe the IOMMU even if it has already been
initialized before booting Linux. In this case during the probe we will
detect that IOMMU is already enabled and we will update the software
cache according to entries that are already registered. Note that we
don't support the refill since pages are not marked as pinned.

Test Plan: Will be tested when L2 will be available

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D1343