k1c: Add support to add ltlb entries
Summary:
In order to add ltlb entries easily, add a ltlb entries allocator to
easily add entries. This allocator simply uses a bitmap and an array to
store virtual address in order to match them when unmapping.

Test Plan: Lock a tlb entry using this function and displayed the ltlb content

Reviewers: O51 Linux Coolidge, gthouvenin

Reviewed By: O51 Linux Coolidge, gthouvenin

Subscribers: alfred, gthouvenin, #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D992