k1c: Modify clear_jtlb_entry to clear ltlb entries
Summary:
In order to be able to remove a ltlb mapping after failure of some init
function (see l2$), we need to be able to remove an LTLB entry.
local_flush_tlb_kernel_range is called after fixmap removal and if such
fixmap was added in the ltlb, then, it must also work for ltlb entry
clearing. To preserve original functionality of panicing when removing
an LTLB entry, add a parameter to specify if only entries from jtlb
should be cleared.

Ref T9042

Test Plan: Verified that a LTLB entry is correctly removed

Reviewers: O51 Linux Coolidge, ysionneau, gthouvenin

Reviewed By: O51 Linux Coolidge, ysionneau, gthouvenin

Subscribers: gthouvenin, alfred, #linux_coolidge_cc

Maniphest Tasks: T9042

Differential Revision: https://phab.kalray.eu/D1311