k1c: Add L2 cache driver
Summary:
In order to correctly handle the L2 cache and request writebacks, we need
to communicate with it. This requires a few modifications to be able to
configure and retrieve informations on the L2 cache.
First of all, the device tree is now passed at boot time to the L2$ firmware.
This is done by passing the DT address through $r1 and the register zone through
$r0. Once the firmware is started, it will setup it status to ready. The driver
will wait for this bit to start and get the mailbox address used to communicate
with it through registers.

Test Plan: Run using SMP + devices

Reviewers: O51 Linux Coolidge, gthouvenin

Reviewed By: O51 Linux Coolidge, gthouvenin

Subscribers: pgmassas, ysionneau, alfred, vchardon, gthouvenin, #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D1012