k1c: dma-noc-test: Fix DMA sync related problems
Summary:
Fix wrong direction set when mapping buffer (invalid FROM/TO device)
Fix wrong access to mapped array after mapping it which an invalid
since the buffer is owned by device once mapped. It must be accessed
first and the map will do a flush of the buffer.
Moreover, fix a problem with buffer ownership before comparing buffers
since the buffer is owned by the device and should be synced for cpu.
Finally, rename the buffer in the cmp routine to make it clear which
buffer is rx and which one is tx.

Test Plan: Run with new L2 cache driver.

Reviewers: O39 ethernet, O51 Linux Coolidge, tcostis, ysionneau

Reviewed By: O39 ethernet, O51 Linux Coolidge, tcostis, ysionneau

Subscribers: gthouvenin, alfred, bmugnier, ysionneau, #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D1391