k1c: Add a fast implementation of strlen
Summary:
This implementation of strlen only does aligned memory accesses.
Since we don't know the total length we do double word loads aligned
on the lower 8-bytes boundary. By doing so we will only cross a memory
page when the string cross it as well. The implementation prologue
takes care of misalignment and realign the input on a 8-bytes boundary.

Test Plan: Run linux with this implementation.

Reviewers: O51 Linux Coolidge, vchardon, cleger, ysionneau

Reviewed By: O51 Linux Coolidge, vchardon, cleger, ysionneau

Subscribers: ysionneau, #linux_coolidge_cc, cleger, vchardon, alfred

Differential Revision: https://phab.kalray.eu/D492