k1c: Split update_mmu_cache
Summary:
When handling vmalloc fault, we end up calling update_mmu_cache
with NULL value. This was done to update the JTLB cache in order to
insert the entry that missed. However, the vma is used leading to a
crash.
Since update_mmu_cache does some checks and then refill the JTLB,
split it and extract the refill part in another function. Ths function
is now called directly in handle_vmalloc_fault to insert the faulting
entry.

Test Plan: Run it with 5.2 kernel in smp mode.

Reviewers: O51 Linux Coolidge, gthouvenin

Reviewed By: O51 Linux Coolidge, gthouvenin

Subscribers: gthouvenin, alfred, #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D1564