k1c: Fix wrong way insertion in $mmc
Summary:
When inserting the JTLB way, we use the value coming from the
current_set_way variable. For simplicity purposes, this value is
incremented monotonically. However, the mask used for insertion is
common to the LTLB and the JTLB which is wrong since LTLB has 16 ways
and JTLB 4. Restrict this mask to correctly insert the way by truncating
it to 2 bits.

Test Plan: Enable the mmc error checker in config and check that it is not triggered.

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D1609