k1c: dma-noc: Probe uc ids from device tree
Summary:
This patch goes read dma-fw-ids from dt to allocate a pool of ids then used to
identify microcode in dma scratchpad. This is necessary to not overlap with l2cache
and for easier configuration.

Test Plan:
Probing and test ok
Messed with dt to test error cases

Reviewers: O39 ethernet, O51 Linux Coolidge, cleger, pkerbrat, tcostis

Reviewed By: O39 ethernet, O51 Linux Coolidge, cleger, tcostis

Subscribers: alfred, pkerbrat, cleger, tcostis, #linux_coolidge_cc

Maniphest Tasks: T9689

Differential Revision: https://phab.kalray.eu/D1100

Conflicts:
	Documentation/devicetree/bindings/dma/kalray,k1c-dma-noc.txt
	drivers/dma/k1/k1c-dma-ucode.c
	drivers/dma/k1/k1c-dma.c
	drivers/dma/k1/k1c-dma.h