k1c: Fix HugePage pte array leak.
Summary:
Fixes the huge_pte_offset in case of vma with
huge tlb page size > 64KB and returns the pmd
(since we have no pud) when asked to allocate
a pte for the same vmas (> 64KB).

Ref: T10728

Test Plan:
Checked that coolval with huge pages
and fork interleaved still runs correctly.

Signed-off-by: Jean-Christophe Pince <jcpince@kalray.eu>

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D2014