k1c: Add ioremap_nocache support
Summary:
ioremap_nocache is currently mapped on mmu DEVICE policy. While
being uncached, this mmu mapping does not allow misaligned accesses.
This can be problematic when simply remapping memory. For memory, we
only need to map it uncached and allow misaligned store (which is legal
though). In order to do so, modify ioremap to accept one additionnal
argument which is the pgprot policy. With this, the ioremap is now a
macro using the old default policy (DEVICE) and for ioremap_nocache,
use UNCACHED policy. This allows to install uncached mapings with
misalignment support.

Test Plan:
Dump the page table using mppa-dump-tlb to check that mapping are
correctly installed using UNCACHED mapping. Moreover, previously
trapping code should not trap anymore.

Reviewers: O51 Linux Coolidge, gthouvenin

Reviewed By: O51 Linux Coolidge, gthouvenin

Subscribers: alfred, gthouvenin, #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D1996