k1c: Pass arguments from RM to PE
Summary:
When bootloader start Linux, the first processor started is RM. This one
receives the arguments through registers and then boots the PE. However,
the PE expects to receive the arguments into its registers r0 -> r3.
We also need to handle the case where a hypervisor start us directly on
PE0. For that purpose, a small wrapper is added and the reset PC of the
PE is set to this address by the RM in case we are started from Linux.
If started from another hypervisor, then, we will boot as usual,
starting from k1c_start.

Test Plan: Checked with debugger that the correct path (pe_start wrapper) is taken.

Reviewers: O51 Linux Coolidge, ysionneau

Reviewed By: O51 Linux Coolidge, ysionneau

Subscribers: ysionneau, alfred, #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D1776