k1c: Add mmu statistic mecanism
Summary:
Profiling MMU TLB refills is needed to analyze the effect of various
modifications on the kernel. In order to do so, instrument the refill
assembly to track duration (avg/min/max) and count of NOMAPPING handling.
These metrics can be displayed using a new debugfs entry. To be
accurate, they are logged percpu and output the following:
```
$ cat /sys/kernel/debug/k1c_mmu_debug/mmu_stats
- Total cycles elapsed: 12600241796
 - CPU 0
  - User refill stats:
   - count: 270410
   - min: 90
   - avg: 332
   - max: 3036
  - Kernel refill stats:
   - count: 4861
   - min: 90
   - avg: 238
   - max: 4442
  - Kernel Direct refill stats:
   - count: 0
   - min: 0
   - avg: 0
   - max: 0
  - Average cycles between refill: 45525
  - tlb_flush_all calls: 5
  - Efficiency: 99%
```

They are categorized in 3 types of refill which are user, kernel direct
mapping and kernel vmalloc mappings.
Number of calls to tlb_flush_all are also tracked.
This patch is a preliminary work to full kernel pagination using correct
R/W/X permissions on kernel pages.

Ref T10849

Test Plan:
Displayed the sysfs and verified the value were coherent.
Also interrupt the refill routine and checked that value are more or less the
expected ones.

Reviewers: O51 Linux Coolidge, gthouvenin

Reviewed By: O51 Linux Coolidge, gthouvenin

Subscribers: gthouvenin, alfred, #linux_coolidge_cc

Maniphest Tasks: T10849

Differential Revision: https://phab.kalray.eu/D2097