k1c: Save $tel, $teh and $mmc when entering refill
Summary:
Since we are going to support full kernel pagination (aka RWX), some
problems that were not visible with always mapped kernel can happen.
For instance, assume the following sequence:
```
  set $tel = $r0
  ;;
a:
  set $teh = $r0
  ;;
  tlbwrite
  ;;
```
Then, if @a is aligned on a page boundary, kernel can caught a trap
NOMAPPING when accessing it. This will trigger a refill which will set
$tel and $teh. When returningi to 'a', $tel value will be overwritten with
the one set in the refill handler leading to unpredictable behavior.
In order to fix that, save $tel and $teh when entering refill handler.
Moreover, since we now have one more free register, use it to store $sr
for registers restoration.

Ref T10856

Test Plan: Run full linux validation

Reviewers: O51 Linux Coolidge, gthouvenin

Reviewed By: O51 Linux Coolidge, gthouvenin

Subscribers: ysionneau, gthouvenin, alfred, #linux_coolidge_cc

Maniphest Tasks: T10856

Differential Revision: https://phab.kalray.eu/D2158