k1c: convert return_from_exception macro to a routine
Summary:
return_from_exception can be factorized since it is always the same
usage everywhere (ie at end of exception) and with a rfe after.
rfe is now moved inside the routine and allows to have a common function
for all exception exit (except for syscalls). This allows to make some
tail calls and reduce the amount of code in the exception page.

Ref T9578

Test Plan: Checked that execution is ok with valid images

Reviewers: O51 Linux Coolidge, ysionneau, gthouvenin

Reviewed By: O51 Linux Coolidge, ysionneau, gthouvenin

Subscribers: ysionneau, alfred, #linux_coolidge_cc

Maniphest Tasks: T9578

Differential Revision: https://phab.kalray.eu/D2337