k1c: Pass registers end address as argument for rm firmware
Summary:
Since we added a check for registers size in l2_firmware, we now need to
pass registers end address as second parameter. Prepare this address and give
it to firmware via $r1.

Depends on D2876

Test Plan: Give a size which is less than expected reg size to check L2 fails to start.

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D2881