k1c: fix issue in protection attributes
Summary:
When running ltp/mprotect04 testcase it hanged. The issue was that the
test change the rights of page from not executable to executable. Our
macro PAGE_EXEC in pgtable.h didn't correspond to a valid hardware
protection attribute and so when a page fault occurs the new TLB entry
was set with NA/NA and the process got a protection trap. This patch
fixed the protection attributes by adding the READ attribute when the
EXEC attribute is used. It also does some refactoring of protection
attributes.

Ref T11492

Test Plan: Pass ltp/mprotect04

Reviewers: O51 Linux Coolidge!

Subscribers: cleger, alfred, #linux_coolidge_cc

Maniphest Tasks: T11492

Differential Revision: https://phab.kalray.eu/D2926