k1c: replace check coherency and check ignored bits
Summary:
Instead of checking the coherency between the SW cache and the TLB when
debug is activated we now check that the write has been done by reading
the new value. We also add a check to ensure that there is no bit in TEL
or TEH that will be ignored. If it is the case we return an error. It is
the case if someone wants to map physical addresses that are too high in
the memory space.

Ref T11389

Test Plan:
Use the test of the dma noc that failed (because it requests a too high
physical address) and ensure that we detect the bit in the bit ignored area.

Reviewers: O51 Linux Coolidge, cleger

Reviewed By: O51 Linux Coolidge, cleger

Subscribers: cleger, #linux_coolidge_cc

Maniphest Tasks: T11389

Differential Revision: https://phab.kalray.eu/D2794