k1c: Preserve bit PAGE_PRESENT accross pte_modify()
Summary:
If we don't preserve this bit then during the cleanup of PTE it can
happens that the PTE is considered as referencing a page that has been
swapped. This wrong assumption introduce the bug in thp03 that was not
related to huge page but to the modification of the PTE when calling
mprotect with NA/NA rights.

Ref T11408

Test Plan: Pass the test thp03

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T11408

Differential Revision: https://phab.kalray.eu/D3018