k1c: Remap text address to writable before patching
Summary:
Now that we have a read only text, text can't be patched directly or it
will cause a PROTECTION trap. To avoid that, remap the text using a
fixmap. This allows to use that function really early in the process
before even setting up paging completely.
This will be used to patch the nomapping handler to skip the performance.

Test Plan:
Using gdb, check that the code was correctly updated.
Typically, a nomapping is triggered to access the fixmap which will be
resolved and then the code will be updated accordingly.

Reviewers: O51 Linux Coolidge, vchardon, gthouvenin

Reviewed By: O51 Linux Coolidge, vchardon, gthouvenin

Subscribers: gthouvenin, vchardon, alfred, #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D2391