pcie_rc: k1c: controller: Support K200 board
Summary:
This version is backward compatible and work on FPGA as well.
The PHY initialisation should be done by the FSBL but for test or debug
purpose it is possible to override FSBL behaviour and to force a PHY reset.
Several root complex may be enabled.

Ref: T11613

Test Plan:
This driver shall work on FPGA
This driver shall work on K200
Several root complex should be supported

Reviewers: O82 pcie root complex!, O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D3350