k1c: IOMMU: Use DROP behavior instead of STALL
Summary:
The IOMMU on the cluster is shared between L2$ and other devices. If the
IOMMU is stalled we won't be able to deal with a L2 cache miss. So to
avoid this we use the DROP mode of the IOMMU and so the IOMMU will be
able to serve requests.

Ref T11757

Test Plan: Check in the dmesg that DROP mode is used.

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T11757

Differential Revision: https://phab.kalray.eu/D3383