k1c: Fix smp boot-time trap handling
Summary:
When using at least 4G of memory, the smp stacks are allocated above
the 512Mb page covered by the LTLB entry 0. This caused a trap which was
handled by the early_trap_handler which simply loops.
However, the whole nomapping handling is ready and would be setup in
setup_processor which correctly install the full exception handling.
This is too far since the stack needs to be mapped and accessing it
triggers a nomapping so install it right before jumping to C code in
smp_secondary_start.
In the mean time, remove useless code which clears ET and HTD in C since
this was halfway done in assembly. Add missing HTD clear in assembly to
allow traps to be taken.

Ref T11777

Test Plan: Test with more than 512MB DDR

Reviewers: O51 Linux Coolidge, gthouvenin

Reviewed By: O51 Linux Coolidge, gthouvenin

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T11777

Differential Revision: https://phab.kalray.eu/D3455