k1c: Introduce break hooks and convert BUG
Summary:
On k1c, we currently rely on errop with a particular immediate to catch BUG.
In near future, we will have KGDB and kprobes support which will use the
same mecanism.
Using set $vsfr is currently used by breakpoints and is a bit more
reliable than using an errop so the break hook mecanism is based on
that rather than errop.
To make this mecanism easily available for such handlers, introduce
break hook mecanism which allows to register handler for break. This
will also be adapted to V2 when break instruction will be reintroduced.
This mecanism is similar to the one available on arm64 but without
different method for each break hook. A single break hook registration
function is available and will add the hook correctly depending on the
type of break hook. To differentiate the cause of break, we use the
source register of the set operation.
Once triggered, break_hook_handler is now invoked in order to
distinguish the cause of the break. If a hook is detected for the break,
then, hook handler will be called. BUG and breakpoint implementation
have been modified to use break hooks instead of previous custom mecanism.

Ref T10482, T11647

Test Plan: Add A BUG_ON in some code and checked that it works as expected.

Reviewers: O51 Linux Coolidge, gthouvenin

Reviewed By: O51 Linux Coolidge, gthouvenin

Subscribers: gthouvenin, #linux_coolidge_cc

Maniphest Tasks: T11647, T10482

Differential Revision: https://phab.kalray.eu/D3150