k1c: Disable hardware loops when restoring $lc,$le,$ls
Summary:
When restoring $le and $ls, if $pc is matching $le then, the core will
jump to $ls to execute the hardware loop. This can be exploited by a
malicious user to crash the kernel or executed unintended code in kernel
space.

Test Plan:
N/A. It would require a special test knowing the return address of syscall
to inject a $le = syscall handler end address and ls = some kernel address.

Reviewers: O51 Linux Coolidge, gthouvenin

Reviewed By: O51 Linux Coolidge, gthouvenin

Subscribers: alfred, #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D3265