Fix issue in pmd_present
Summary:
If PMD has its value with all lower 32bits set to 0 then pmd_present
always reported false. As it is used to know if it has been swapped the
kernel though that the PMD was swapped. As we don't support swap it
retries for ever. This issue happens for example when a page is mapped
on physical address 0x100000000. Fixes PUD at the same time.

Ref T11487

Test Plan: This should fix the issue when perf tests hang

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T11487

Differential Revision: https://phab.kalray.eu/D4265