k1c: eth: Set retimer speed from ethernet driver
Summary:
The old way to set retimer speed is to use reg-init property in device tree to
write proper i2c sequence. This is static and duplicated information among
retimer node and ethernet node.
This commits sets the retimers speed from ethernet driver using its own speed.
This will also be handy for autoneg as we will need to reconfigure retimers on
the fly.
Reg-init function is kept as it could be used for other purposes.
Note that I2C bus is locked in i2c transfer and no explicit lock is needed.

Test Plan: No regressions

Reviewers: O39 ethernet, O51 Linux Coolidge, tcostis, cleger

Reviewed By: O39 ethernet, tcostis

Subscribers: cleger, tcostis, #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D4057