k1c: Add lx-k1c-virt-to-phys command
Summary:
This command allows to find a physical address matching a virtual one by
walking the page table. It requires at least 1 argument which is the
virtual to be translated. If given, the second one is the pgd to
consider when walking the page table.
If a clean entry is found, then the output will be the following:
```
Trying to find phys_address for 0xffffff8000655588
PFN: 0x0000000100650000, PS: 64 Ko, bits: H G W R CP P
Physical address: 0x0000000100655588
```

Test Plan: Tried to display various addresses using this script

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D4290