pcie_rc: k1c: dynamically deduce number of lane from nfurcation
Summary:
The "kalray,nb_lane" field in device tree is no longer mandatory.
If not present the number of lane will be deduced from the nfurcation.
It is still possible however to define this field in order to define a smaller value.

Ref: T12232

Test Plan:
Confirm in nfurcation 1X16, 2X8, 4X4, 8X2 that the number of lane is correctly deduced.
Ensure that those settings can be overriden to smaller value.

Reviewers: O82 pcie root complex!, O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D4358