k1c: Use remaining space for vmalloc
Summary:
Previously, we were using a fixed size vmalloc space. This revealed to
be too tight when using systems with a big number of devices. In order
to be more flexible, simply use the whole remaining memory starting from
kernel direct memory base up to fixmap start. While doing this change,
add a guard of 1G to allow catching erroneous accesses. With the current
setup, this lead to ~445G of vmalloc space.

Test Plan:
Run linux valdiation (and particularly modules) and enumerate lot of
PCIe devices.

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D4398