pcie_rc: k1c: support 32 bit MSI
Summary:
This commit force the MSI write destination to be located in 32 bits DDR
alias. This is required in particular when the iommu is not active in
order to support 32 bits MSI. The device tree shall be updated accordingly
in order to define the 32 bits aliased ddr location.
The memory relocation is required only for the MSI controller.
The device tree should be defined as follow :

    pcimsi_ctrl: pcimsi {
	/* Translate 64 bits RAM addresses to 32 bits RAM alias */
	dma-ranges = <0x0 0x80000000 0x1 0x00000000 0x0 0x80000000>;
    };

Ref: T12331

Test Plan:
Test with hotplug and config device state change reception. It shall work
in the following case :
- IOMMU not activated and device tree defined as above
- IOMMU activated and device tree defined as above
- No offset shall be applied when dma-ranges entry is not present in
  the device tree

Reviewers: O82 pcie root complex!, O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D4498