iommu: k1c: Properly manage group reference count
Summary:
In the context of PCIe, the code is being exercised in two new
ways. Several devices can be added in the same group, when hotplug is
being used a device can be dynamically added and removed.
This commit increase the group reference count each time a device is
added in a group. As there more than one device in group, do not
systematically destroy the group when a device is remove, instead
decrement the group reference count and destroy it only when this
reference count is 0.

Ref: T12330

Test Plan:
Ensure hot plug is working when :
- the IOMMU is disabled
- the IOMMU is enabled

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D4506