k1c: double the size of the kernel stack
Summary:
When using perf the size of the stack used for the kernel can be bigger
than 8Ko. So this patch increases the size to 16Ko. It is because
function used by perf are storing data on the stack and most of them
are using around 1Ko of stack. This is because perf structure
contains pt_regs and on k1, regs are requiring 596 bytes.

Ref T12253

Test Plan: Run perf manually in a loop to check that the bug reported in T12253 is fixed

Reviewers: O51 Linux Coolidge, cleger

Subscribers: cleger, #linux_coolidge_cc

Maniphest Tasks: T12253

Differential Revision: https://phab.kalray.eu/D4520