pcie_rc: k1c: Allow a user to downgrade max link speed
Summary:
The root complex support up to GEN4 link speed. Some board may have been
routed with lower specification. For example an endpoint might be enumerated
as GEN3 but some AER errors might be reported from time to times. In
that case this commit allow a user to use a smaller generation.
The configuration is done from the device tree using the standard pci
settings "max-link-speed".

Test Plan: Confirm a SSD can be enumerated in GEN3, GEN2 and GEN1

Reviewers: O82 pcie root complex!, O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D4601