k1c: eth: Add implicit priorities to parsers based on OSI layer
Summary:
Previously if two parsers match at the same time the resulting matching parser is unknown.
We don't have the ability to set parser priorities using ethtool, so we make the
assumption the user wants the more specific parser to be elected.
This patch uses the OSI layer as a priority for the hardware parser, resulting
in the higher layer rule being elected.

Test Plan:
Tested by DCBU with a simple mac rule matching all traffic and a specific TCP
one.

Reviewers: O39 ethernet!, O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D4636