rpmsg: char: use cdev_device_add/del
Summary:
In rpmsg_char driver, cdev and dev have the same lifecycle. In the
device release, the structure containing the cdev is freed. However, the
cdev is not mandatorily freed on the cdev_del but likely to be free
later when the cdev refcount is zero (in __fput) which lead to using a
freed object. In order to fix that, use cdev_device_add which will free
the dev only when the cdev is released, leading to objects being kfreed
after all references have been dropped.

Ref T12417

Test Plan: Run rproc tests on linux with PAGE_POISONING and DEBUG_SLUB enabled.

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T12417

Differential Revision: https://phab.kalray.eu/D4648