remoteproc: k1c: Add support for multiple memory-regions
Summary:
Some remote firmwares will require to be loaded in DDR. To do so, allow
memory-region to take multiple regions and ioremap them in the driver.
Currently, this does not allow allocating vdev buffers in this zone
anymore but this was never used yet. This support will be added in a
separate commit to match with ClusterOS configuration.
To add a reserved memory for clusters, the following should be added in
the device tree:
```
reserved-memory {
	cluster_mem: mem@110000000{
		reg = <0x1 0x10000000 0x0 0x100000>;
		no-map;
	};
}

cluster1 {
	memory-region = <&cluster_mem>;
}

```

Ref T11923

Test Plan: A special ddr test has been added in rproc tests

Reviewers: O51 Linux Coolidge, vchardon

Reviewed By: O51 Linux Coolidge, vchardon

Subscribers: vchardon, gthouvenin, sjones, #linux_coolidge_cc

Maniphest Tasks: T11923

Differential Revision: https://phab.kalray.eu/D4080