spi: dw: add support for spi-mem
Summary:
This commit adds support for spi-mem operation. While the generic support
could work for other controllers, the dw-ahb-ssi controller does not behave
like previous controllers regarding the chip select.
Indeed, once selected the controller won't assert it immediatly but only
when sending data. In standard SPI mode, the controller will deassert the
chip select if TX fifo is empty or RX fifo is full. This is a problem for
the SPI-MEM transfers which expect the transfer to be done in only one
transaction (ie slave selected during the whole transfer). With the current
support via spi-mem generic operation, the operation are splitted. cmd is
sent using a single write, then addr and finally data. This result in
multiples transactions from the flash POV and thus are non working.
In order to guarantee that operation will always be atomic and that chip
select is always selected for the whole transaction we adjust the size of
operations to a maximum transfer size equal to the fifo_len. If using a
larger transfer size, then it would mean that we must ensure that the
interrupt happens soon enought to refill TX fifo or pop the RX fifo.

Depends On D5491

Test Plan: Execute linux and run tests with mtd_debug command and logic analyzer

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D5498