spi: dw: add enhanced SPI support in spi-mem
Summary:
dw-ahb-ssi controller supports an enhanced SPI mode. This mode allows to
driver the SPI in dual/quad/octal and has a special behavior regarding
the chip select if clock stretching is enabled. This mode allows to do
up to 64k transfer in one shot without letting the chipselect be
deselected due to host not keeping up the rate of the TX fifo or RX
fifo. This mode requires the number of expected data frames to be
written in CTRL1 register which will indicate how long the controller
will keep the slave selected. Once done, even if the RX fifo is full
or if the TX fifo is empty, the controller will keep the chip select
asserted.

Depends On D5498

Test Plan:
Execute on kvx and check with mtd_debug that the transfer is done using
quad mode.

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D5500