kvx: rework break_hook handlers
Summary:
Previously, ea register was passed but completely not relevant since
when using vsfr fault, ea is not set. In order to make that more clean,
remove also es since it is obtained in regs and likely not to be used.
To allow break hook handlers to distinguish between multiple break hook
causes, pass the break_hook struct that caused the break in parameters.

Depends On D5334

Test Plan: Checked that BUG framework still works as expected

Reviewers: O51 Linux Coolidge, gthouvenin

Reviewed By: O51 Linux Coolidge, gthouvenin

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D5335