watchdog: kvx_wdt: add wdt_in_kernel module parameter
Summary:
Currently, when loaded, the driver will immediatly enable the watchdog
in kernel. While being useful, this does not allow using KGDB or the
watchdog will reboot the processor if not fed. To allow using KGDB, add
an option to disable the watchdog in kernel (wdt_in_kernel) which allows
disabling this functionality. When set to 0, the watchdog will only be
start when the userspace opens /dev/watchdog.

Depends On D5338

Test Plan:
Booted kernel with kvx_wdt.wdt_in_kernel disabled and check via gdb that
watchdog is not started.

Reviewers: O51 Linux Coolidge, gthouvenin

Reviewed By: O51 Linux Coolidge, gthouvenin

Subscribers: gthouvenin, #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D5339