kvx: add early param to disable streaming
Summary:
When debugging, it is often useful to disable streaming to get a precise
trap instead of an interrupt. Add an early param use_streaming which
allows to disable use at boot time. USE field is both cleared in the
early param handler for the first booting processor and then other
processors in SMP will disable it in setup_processor.

Ref T12794

Test Plan: Start kernel with kvx.streaming=off and check that USE is disabled.

Reviewers: O51 Linux Coolidge, gthouvenin

Reviewed By: O51 Linux Coolidge, gthouvenin

Subscribers: gthouvenin, #linux_coolidge_cc

Maniphest Tasks: T12794

Differential Revision: https://phab.kalray.eu/D5445