kvx: align pt_regs struct size on STACK_ALIGNMENT
Summary:
In order to have efficient lo/so on the stack, we need to align the
stack on 256 bits. This commit adds padding in the pt_regs struct to
always guarantee that adding pt_regs on stack will keep the stack
aligned. Moreover, this allows to keep the task_pt_regs macro the same
without having to align the stack pointer.

Depends On D5672

Test Plan:
Execute kernel with userspace to guarantee everything still works as
expected.

Reviewers: O51 Linux Coolidge, ysionneau

Reviewed By: O51 Linux Coolidge, ysionneau

Subscribers: #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D5673