kvx: add debug hook
Summary:
In order to allow some services to request single stepping, cleanup the
existing debug entry to allow having multiple debug requester. This is
based of break hook mecanism and multiple step handlers can be requested.
This allows to hook KGDB support onto it.

Depends On D5339

Test Plan:
Compile and check with KGDB that the step is correctly catched and
served.

Reviewers: O51 Linux Coolidge, gthouvenin

Reviewed By: O51 Linux Coolidge, gthouvenin

Subscribers: gthouvenin, #linux_coolidge_cc

Differential Revision: https://phab.kalray.eu/D5345